fix: validate resource kind.
In order to allow only pod resource in the template policy, it's
necessary to check the resource kind before trying to parse. This is
necessary because the parse function can be able to parse it even if the
resource is not pod. This is necessary to improve the docs about writing
policy in rust.

Signed-off-by: José Guilherme Vanz <jguilhermevanz@suse.com>