refactor required to fix online tutorial
This changes the way the settings class is built inside of the tests.
That's required to ensure the online tutorial works as expected.

The tutorial changes the `Settings` class to have a `HashSet` attribute.
This causes the invocations of `let s Settings = Settings {}` to fail
because the new attribute is not specified.

Moving to use the `Default::default` contructor fixes everything.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>